# specs/openapi/propnexus.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "PropNexus API",
    "version": "1.0.0",
    "description": "Generado desde el router montado (`pnpm --filter @plataforma/api docs:openapi`), no mantenido a mano — ver apps/api/scripts/generate-openapi.ts. El código de éxito se lee del `res.status(2xx)` real del handler (con el de creación si el handler tiene más de uno, caso idempotente); las respuestas no traen schema (ver el comentario del generador). Body y query sí son el schema Zod real donde existe."
  },
  "servers": [
    {
      "url": "http://localhost:3001/api/v1",
      "description": "Local (pnpm dev)"
    }
  ],
  "paths": {
...
    "/api/v1/projects": {
      "get": {
        "summary": "GET /api/v1/projects",
        "description": "auth + autoriza(rol(admin|developer|buyer|verifier|notary) · scope(projectScope(cualquier membresía)))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "query",
            "name": "status",
            "schema": {
              "type": "string",
              "enum": [
                "planning",
                "in_progress",
                "delayed",
                "completed"
              ]
            }
          },
          {
            "in": "query",
            "name": "city",
            "schema": {
              "type": "string",
              "minLength": 1
            }
          },
          {
            "in": "query",
            "name": "q",
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 120
            }
          },
          {
            "in": "query",
            "name": "sort",
            "schema": {
              "type": "string",
              "enum": [
                "recent",
                "name",
                "delivery"
              ]
            }
          },
          {
            "in": "query",
            "name": "bbox",
            "schema": {
              "type": "string",
              "pattern": "^-?\\d+(\\.\\d+)?(,-?\\d+(\\.\\d+)?){3}$"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Body o query no pasan el schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      },
      "post": {
        "summary": "POST /api/v1/projects",
        "description": "auth + autoriza(rol(admin) · soloRol)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "totalUnits": {
                    "default": 0,
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "estimatedDelivery": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  },
                  "status": {
                    "default": "planning",
                    "type": "string",
                    "enum": [
                      "planning",
                      "in_progress",
                      "delayed",
                      "completed"
                    ]
                  }
                },
                "required": [
                  "name",
                  "slug"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "400": {
            "description": "Body o query no pasan el schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/projects/{id}": {
      "get": {
        "summary": "GET /api/v1/projects/:id",
        "description": "auth + autoriza(rol(admin|developer|buyer|verifier|notary) · proyecto(id → developer|buyer|verifier))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      },
      "patch": {
        "summary": "PATCH /api/v1/projects/:id",
        "description": "auth + autoriza(rol(admin) · soloRol)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "latitude": {
                    "type": "number"
                  },
                  "longitude": {
                    "type": "number"
                  },
                  "totalUnits": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "estimatedDelivery": {
                    "type": "string",
                    "format": "date-time",
                    "pattern": "^(?:(?:\\d\\d[2468][048]|\\d\\d[13579][26]|\\d\\d0[48]|[02468][048]00|[13579][26]00)-02-29|\\d{4}-(?:(?:0[13578]|1[02])-(?:0[1-9]|[12]\\d|3[01])|(?:0[469]|11)-(?:0[1-9]|[12]\\d|30)|(?:02)-(?:0[1-9]|1\\d|2[0-8])))T(?:(?:[01]\\d|2[0-3]):[0-5]\\d(?::[0-5]\\d(?:\\.\\d+)?)?(?:Z))$"
                  },
                  "status": {
                    "type": "string",
                    "enum": [
                      "planning",
                      "in_progress",
                      "delayed",
                      "completed"
                    ]
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          },
          "400": {
            "description": "Body o query no pasan el schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      },
      "delete": {
        "summary": "DELETE /api/v1/projects/:id",
        "description": "auth + autoriza(rol(admin) · soloRol)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "204": {
            "description": "Sin contenido"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
...
    "/api/v1/projects/{id}/documents": {
      "get": {
        "summary": "GET /api/v1/projects/:id/documents",
        "description": "auth + autoriza(rol(admin|developer|buyer|verifier|notary) · proyecto(id → developer|buyer|verifier))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
...
    "/api/v1/developer/projects": {
      "get": {
        "summary": "GET /api/v1/developer/projects",
        "description": "auth + autoriza(rol(admin|developer) · scope(projectScope(developer)))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      },
      "post": {
        "summary": "POST /api/v1/developer/projects",
        "description": "auth + autoriza(rol(admin|developer) · soloRol)",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "name": {
                    "type": "string",
                    "minLength": 1
                  },
                  "slug": {
                    "type": "string",
                    "minLength": 1
                  },
                  "address": {
                    "type": "string"
                  },
                  "city": {
                    "type": "string"
                  },
                  "country": {
                    "type": "string"
                  },
                  "totalUnits": {
                    "type": "integer",
                    "minimum": 0,
                    "maximum": 9007199254740991
                  },
                  "estimatedDelivery": {
                    "type": "string"
                  }
                },
                "required": [
                  "name",
                  "slug"
                ],
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "OK"
          },
          "400": {
            "description": "Body o query no pasan el schema",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationError"
                }
              }
            }
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/developer/projects/{id}": {
      "get": {
        "summary": "GET /api/v1/developer/projects/:id",
        "description": "auth + autoriza(rol(admin|developer) · proyecto(id → developer))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "schema": {
              "type": "string",
              "pattern": "^[a-z][a-z0-9]{23}$"
            },
            "required": true
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
    "/api/v1/developer/progress": {
      "get": {
        "summary": "GET /api/v1/developer/progress",
        "description": "auth + autoriza(rol(admin|developer) · scope(projectScope(developer)))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
...
    "/api/v1/developer/kpis": {
      "get": {
        "summary": "GET /api/v1/developer/kpis",
        "description": "auth + autoriza(rol(admin|developer) · scope(projectScope(cualquier membresía)))",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "401": {
            "description": "Sin sesión o token inválido"
          },
          "403": {
            "description": "Rol o membresía insuficiente"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ValidationError": {
        "type": "object",
        "properties": {
          "formErrors": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "fieldErrors": {
            "type": "object",
            "propertyNames": {
              "type": "string"
            },
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            }
          }
        },
        "required": [
          "formErrors",
          "fieldErrors"
        ],
        "additionalProperties": false
      }
    },
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}